Erweiterung der Praezision auf 8 Nachkommastellen (dbversion 19) (#156)
* Erweiterung der Praezision auf 8 Nachkommastellen (dbversion 19)

Betraege, Kurse, Gebuehren und Steuern werden jetzt mit decimal(20,8)
gespeichert (bisher decimal(20,6), kursevent.value sogar nur decimal(10,5)).

- Changeset v19 datenbankspezifisch (H2: ALTER COLUMN, MySQL: MODIFY COLUMN),
  da H2 2.x kein MODIFY COLUMN mehr versteht
- Kursberechnung beim CSV-Import: Scale 5 -> 8
- Umsatz-Editor: Eingabefelder zeigen/erhalten bis zu 8 Nachkommastellen
- database.md: Analyse aller Praezisionsgrenzen in Schema, Backend und GUI
- PrecisionTest: Roundtrip-Tests (gespeichert == gelesen) und
  Migrationstest v18 -> v19 mit Erhalt der Altwerte (H2)


* Fix Javac Warning:

    [javac] Warnung: [options] Systemmodulpfad ist nicht zusammen mit -source 17 festgelegt
    [javac]   Wenn Sie den Speicherort der Systemmodule nicht festlegen, kann dies zu Klassendateien führen, die auf JDK 17 nicht ausgeführt werden können
    [javac]     --release 17 wird anstelle von -source 17 -target 17 empfohlen, weil dadurch der Speicherort der Systemmodule automatisch festgelegt wird

* Ändere Ansicht für Anzahl Wertpapiere auf 8 statt 5

* remove unused imports on changed files only. Fix database.md documentation

* Geldbeträge, Kurse und Stückzahlen durchgängig als BigDecimal führen

Seit dbversion 19 liegen alle Geld-/Kursspalten als decimal(20,8) und
Stückzahlen als decimal(20,10) in der Datenbank. Die Schreibpfade haben
diese Präzision bisher nicht durchgehalten: die Werte liefen als double
durch Abruf, Import und GUI und wurden erst beim Speichern zurück nach
BigDecimal gewandelt — teilweise über new BigDecimal(double), das die
Binärdarstellung exakt expandiert statt die dargestellte Dezimalzahl zu
übernehmen. Aus 54321.12345678 wurde so ein Wert mit Stellen, die nie
eingegeben wurden. Damit ist die neue Spaltenpräzision erst ab hier
tatsächlich nutzbar.

Schreib-Fassade und Datenobjekte
- Utils.addUmsatz(...) und Utils.addBestand(...) nehmen anzahl, kurs,
  kosten, wert, gebuehren und steuern als BigDecimal entgegen. Die
  new BigDecimal(double)-Aufrufe im Rumpf entfallen ersatzlos, die
  Vorzeichenprüfung bei Kauf/Verkauf läuft über signum().
- Bestand und Kursevents (rmi + Impl) führen anzahl, kurs, wert und
  value als BigDecimal. BestandImpl.getWert() hatte den aus JDBC
  gelieferten BigDecimal bisher per doubleValue() verworfen.
- createClassesFromSQL erzeugt für decimal-Spalten jetzt BigDecimal
  statt Double — ein (Double)-Cast auf getAttribute() knallt zur
  Laufzeit, weil JDBC BigDecimal liefert.

Aufrufer
- HBCIDepotBestandJob / HBCIDepotUmsatzJob geben die hbci4j-Werte
  (Value.getValue() liefert bereits BigDecimal) unverändert weiter,
  statt sie über doubleValue() zu schleusen.
- HBCIDepotUmsatzJob: der Einzelbetrag aus Gesamtbetrag/Anzahl braucht
  als BigDecimal-Division eine explizite Rundung — scale 8, HALF_UP,
  passend zur Spalte kurs decimal(20,8).
- CortalConsorsMitHBCIJSONWrapper liest Kurs und Stückzahl über
  new BigDecimal(String) statt Double.parseDouble. getKosten() rundet
  mit setScale(2, HALF_EVEN), was das bisherige Verhalten von
  Math.rint() (kaufmännisch zur geraden Ziffer) erhält.
- BestandImportAction, Bestandspruefung und
  UmsatzeAusBestandsAenderung reichen die ohnehin vorhandenen
  BigDecimal-Werte durch, statt sie vorher nach double zu wandeln.
- Utils.getDoubleFromZahl entfällt zugunsten von
  Zahlen.ausDeutscherSchreibweise (BigDecimal statt Double); im
  Produktivcode derzeit ohne Aufrufer, im Test abgedeckt.

GUI
- VarDecimalFormat setzt setParseBigDecimal(true), damit
  DecimalInput.getNumber() BigDecimal liefert.
- UmsatzEditorControl rechnet Kurswert, Gesamtsumme und alle
  gespeicherten Werte in BigDecimal. Die Felder sind auf DecimalInput
  typisiert, damit getNumber() erreichbar ist. Vorbelegung ist null
  statt Double.NaN, Steuern/Gebühren sind mit BigDecimal.ZERO
  vorbelegt und werden beim Speichern auch als 0 gewertet, wenn das
  Feld leer ist. Das Anzahl-Feld zeigt 8 statt 6 optionale
  Nachkommastellen, passend zu decimal(20,10).
- BestandTableControl summiert den Gesamtdepotwert in BigDecimal;
  BestandPieChartControl addiert die Werte gleicher Wertpapiere aus
  verschiedenen Depots exakt und rundet die Label-Anzeige über
  setScale(0, HALF_UP) statt Math.round(double).

Fremde Schnittstellen
- Neu: de.open4me.depot.tools.Zahlen mit toBigDecimal(Object),
  optionalerBetrag(Object) und den beiden OrderID-Berechnungen.
  Bewusst frei von Jameica-Abhängigkeiten, damit die Rechenregeln
  ohne laufende Anwendung testbar bleiben — Settings und Utils
  greifen schon beim Laden der Klasse auf die Application zu.
- TransactionMessageConsumer und PortfolioMessageConsumer wandeln die
  untypisierten QueryMessage-Werte über Zahlen um, statt hart auf
  Double zu casten. Fehlende Gebühren/Steuern werden über
  optionalerBetrag zu 0 — so führt die Datenbank sie seit dbversion 12.
- PortfolioMessageConsumer: doppelter Konto-Import und der ungenutzte
  ApplicationException-Import entfernt.

Bewusst NICHT auf BigDecimal umgestellt
Diese Stellen bleiben double, weil die Gegenseite nichts anderes
anbietet oder eine Änderung Bestandsdaten brechen würde. Sie sind im
Code jeweils kommentiert:

- Zahlen.fuerHash(): die synthetische OrderID geht weiterhin über
  doubleValue(). Anzahl und Kurs waren früher Double, und
  BigDecimal.toString() liefert eine andere Darstellung ("10" statt
  "10.0") und damit einen anderen hashCode. Ohne diese Umrechnung
  würden bereits importierte Orders nicht mehr wiedergefunden und beim
  nächsten Abruf ein zweites Mal angelegt. Betrifft beide Formeln —
  Utils.addUmsatz (orderid == null) und den Umsatz-Editor.
- Bestandsabfragen (AccountBalanceProvider): Hibiscus' Value kennt nur
  double. Statt weiterhin je Wertpapier in double zu addieren, laufen
  die Tagessummen jetzt exakt über ein BigDecimal[] und werden erst
  einmalig am Ende per doubleValue() an Value übergeben — gerundet
  wird damit nur noch an der Systemgrenze.
- Bestandspruefung: Konto.setSaldo(double) ist Hibiscus-API, der Saldo
  wird intern als BigDecimal gerechnet und erst beim Setzen gewandelt.
- UmsatzEditorControl.toDecimal(): der NaN/Infinity-Test läuft über
  doubleValue(), weil ein leeres bzw. kaputtes Eingabefeld genau
  darüber erkennbar ist. Bewusst über DecimalInput.getNumber(), nicht
  getValue() — getValue() wandelt intern immer nach Double.
- VarDecimalFormat betrifft nur parse(), nicht format(). Die
  Anzeigepräzision bleibt verlustbehaftet: was im Eingabefeld steht,
  wird beim Speichern zurückgeschrieben.
- BestandPieChartControl: JFreeChart arbeitet auf Number, die Werte
  werden als BigDecimal übergeben und über Zahlen.toBigDecimal wieder
  ausgelesen; die Diagramm-Darstellung selbst bleibt intern double.
- Die MessageConsumer nehmen weiter Double von fremden Plugins
  entgegen. Ein Wechsel der Queue-Typen würde die Sender brechen —
  deshalb die Konvertierung über toString(), die Double, Long und
  String gleichermaßen verlustfrei übernimmt.

Tests
- Neu: ZahlenTest mit 13 JUnit-4-Tests für die Konvertierungen
  (BigDecimal/Double/Long/String/null), die ZERO-Vorbelegung der
  optionalen Beträge, die deutsche Schreibweise und vor allem die
  Stabilität beider OrderID-Formeln gegenüber dem Stand vor der
  Umstellung. Läuft ohne Jameica.
- Die Speicherpräzision der Spalten selbst deckt weiterhin
  database.PrecisionTest auf JDBC-Ebene ab.

---------

+++ src/de/open4me/depot/tools/Bestandspruefung.java
@@ -136,7 +136,7 @@ public class Bestandspruefung {
 			}
 			
 			// Bestand hinzufügen
-			Utils.addBestand(position.getKey().toString(), k, position.getValue().doubleValue(), kurs.doubleValue(), kursW, wert.doubleValue(), wertW, new Date(), bewertung);
+			Utils.addBestand(position.getKey().toString(), k, position.getValue(), kurs, kursW, wert, wertW, new Date(), bewertung);
 		}
 		k.setSaldo(saldo.doubleValue());
 		k.store();
